+++ tests/multiple-editor/first-editor.spec.ts
@@ -1,5 +1,5 @@
 import './setup-first.js';
-import '../basic.spec.ts?first';
+import '../basic.case.ts?first';
 
 // console.log(test);
 import { setScope } from '../utils/playwright.js';
