+++ tests/multiple-editor/second-editor.spec.ts
@@ -1,5 +1,5 @@
 import './setup-second.js';
-import '../basic.spec.ts?second';
+import '../basic.case.ts?second';
 
 import { setScope } from '../utils/playwright.js';
 setScope('');
